Updated libjpeg to version 9d.

+++ renderer/r_image.c
@@ -1284,8 +1284,8 @@ void jpg_skip_input_data (j_decompress_ptr cinfo, long num_bytes)
 		VID_Printf(PRINT_ALL, "Premature end of JPEG data\n");
 }
 
-//void jpeg_mem_src(j_decompress_ptr cinfo, byte *mem, int len)
-void jpeg_mem_src (j_decompress_ptr cinfo, const byte *mem, unsigned long len)
+//void jpeg_mem_src (j_decompress_ptr cinfo, const byte *mem, unsigned long len)
+void jpeg_mem_src (j_decompress_ptr cinfo, const byte *mem, size_t len)
 {
     cinfo->src = (struct jpeg_source_mgr *)(*cinfo->mem->alloc_small)((j_common_ptr) cinfo, JPOOL_PERMANENT, sizeof(struct jpeg_source_mgr));
     cinfo->src->init_source = jpg_null;
